requirements.txt
click==8.0.0
colorama==0.4.4
dnspython==1.16.0
Flask==2.0.0
Flask-PyMongo==2.3.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
pymongo==3.11.4
Werkzeug==2.0.0
gunicorn
python-dotenv